backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pillow
pytesseract
groq
reportlab
pypdf
